services/retriever_service/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
redis
numpy